requirements-dev.txt
# Development and testing dependencies
-r requirements.txt

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-django==4.5.2
pytest-cov==4.1.0
python-dotenv==1.0.0

# Code quality
flake8==6.1.0
isort==5.13.2
pre-commit==3.5.0
